# swagger/doc.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Backend API for RazorBlog including Authors, Blogs, Comments, Shares",
    "title": "RazorBlog API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "RazorBlog Support",
      "url": "http://razorblog.io",
      "email": "support@razorblog.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/blogs/{blog_id}/shares": {
      "get": {
        "description": "Returns a list of shares for a specific blog post",
        "produces": ["application/json"],
        "tags": ["Shares"],
        "summary": "List shares for a blog",
        "parameters": [
          {
            "type": "string",
            "description": "Blog ID",
            "name": "blog_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Share"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "/shares": {
      "post": {
        "description": "Records a share of a blog post on a platform",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Shares"],
        "summary": "Create a new share",
        "parameters": [
          {
            "description": "Share info (blog_id, platform)",
            "name": "share",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Share"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "models.Share": {
      "type": "object",
      "properties": {
        "blog_id": {
          "description": "Blog being shared",
          "type": "string"
        },
        "created_at": {
          "description": "Timestamp of the share",
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "platform": {
          "description": "e.g., Twitter, Facebook",
          "type": "string"
        }
      }
    }
  }
}